requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
h11==0.13.0
idna==3.3
outcome==1.2.0
Pillow==9.2.0
pycparser==2.21
PySocks==1.7.1
requests==2.28.1
selenium==4.4.3
Send2Trash==1.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
